refactor: 重构所有 LuCI 包，使用 luci.mk 并添加 i18n 支持
主要变更：
- 所有包改用 luci.mk 构建系统
- Lua 源码从 root/usr/lib/lua/luci/ 移动到 luasrc/
- 源码中的中文 msgid 改为英文，添加 po 翻译文件
- luci-app-simple2fa: 实现通用主题兼容方案，通过 CGI 动态注入验证码输入框

涉及包：
- luci-app-cymfrpc
- luci-app-cymfrps
- luci-app-simple2fa
- luci-app-vpnrss
