+++ luci-app-cymfrpc/Makefile
@@ -1,36 +1,23 @@
 include $(TOPDIR)/rules.mk
 
-PKG_NAME:=luci-app-cymfrpc
+LUCI_TITLE:=LuCI Support for CymFrpc
+LUCI_DEPENDS:=+cymfrpc +luci-base +luci-compat
+LUCI_PKGARCH:=all
+
 PKG_VERSION:=1.0.0
 PKG_RELEASE:=1
-PKG_MAINTAINER:=findnr
 
-include $(INCLUDE_DIR)/package.mk
-
-define Package/luci-app-cymfrpc
-  SECTION:=luci
-  CATEGORY:=LuCI
-  SUBMENU:=3. Applications
-  TITLE:=LuCI Support for CymFrpc
-  DEPENDS:=+cymfrpc
-  PKGARCH:=all
-endef
-
-define Package/luci-app-cymfrpc/description
-  LuCI support for CymFrpc (Multi-instance, Raw Config).
-endef
-
-define Build/Compile
-endef
+include $(TOPDIR)/feeds/luci/luci.mk
 
 define Package/luci-app-cymfrpc/install
-	$(INSTALL_DIR) $(1)/etc/config
-	$(INSTALL_DIR) $(1)/etc/init.d
 	$(INSTALL_DIR) $(1)/usr/lib/lua/luci/controller
 	$(INSTALL_DIR) $(1)/usr/lib/lua/luci/model/cbi/cymfrpc
-
-	$(CP) ./root/* $(1)/
-	# Fix CRLF line endings
+	$(INSTALL_DIR) $(1)/etc/config
+	$(INSTALL_DIR) $(1)/etc/init.d
+	
+	$(INSTALL_DATA) ./luasrc/controller/cymfrpc.lua $(1)/usr/lib/lua/luci/controller/
+	$(INSTALL_DATA) ./luasrc/model/cbi/cymfrpc/*.lua $(1)/usr/lib/lua/luci/model/cbi/cymfrpc/
+	$(CP) ./root/etc/* $(1)/etc/
 	sed -i 's/\r//' $(1)/etc/init.d/cymfrpc
 	chmod +x $(1)/etc/init.d/cymfrpc
 endef
