--- luci-app-cymfrpc/root/usr/lib/lua/luci/controller/cymfrpc.lua
@@ -1,10 +1,0 @@
-module("luci.controller.cymfrpc", package.seeall)
-
-function index()
-	if not nixio.fs.access("/etc/config/cymfrpc") then
-		return
-	end
-
-	entry({"admin", "services", "cymfrpc"}, cbi("cymfrpc/overview"), _("CymFrpc 客户端"), 60).dependent = true
-	entry({"admin", "services", "cymfrpc", "instance"}, cbi("cymfrpc/instance"), nil).leaf = true
-end
