--- luci-app-cymfrps/root/usr/lib/lua/luci/controller/cymfrps.lua
@@ -1,10 +1,0 @@
-module("luci.controller.cymfrps", package.seeall)
-
-function index()
-	if not nixio.fs.access("/etc/config/cymfrps") then
-		return
-	end
-
-	entry({"admin", "services", "cymfrps"}, cbi("cymfrps/overview"), _("CymFrps 服务端"), 60).dependent = true
-	entry({"admin", "services", "cymfrps", "instance"}, cbi("cymfrps/instance"), nil).leaf = true
-end
