+++ luci-app-simple2fa/luasrc/view/simple2fa/qrcode_view.htm
