+++ luci-app-simple2fa/po/templates/simple2fa.pot
@@ -1,0 +1,35 @@
+msgid ""
+msgstr "Content-Type: text/plain; charset=UTF-8"
+
+msgid "Two-Factor Auth"
+msgstr ""
+
+msgid "Two-Factor Authentication"
+msgstr ""
+
+msgid "Enable 2FA to protect your router login."
+msgstr ""
+
+msgid "Settings"
+msgstr ""
+
+msgid "Enable Two-Factor Auth"
+msgstr ""
+
+msgid "Secret Key"
+msgstr ""
+
+msgid "If you cannot scan the QR code, enter this key manually."
+msgstr ""
+
+msgid "Refresh Secret"
+msgstr ""
+
+msgid "Warning: After refreshing, you must reconfigure all authenticator apps."
+msgstr ""
+
+msgid "Scan QR Code"
+msgstr ""
+
+msgid "Use Google Authenticator, Authy or Microsoft Auth to scan this QR code."
+msgstr ""
